pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-scanner-mvp"
version = "0.1.0"
description = "MVP scanner for GitHub repositories with optional llama.cpp enrichment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
